Microshift in RHEL9

+++ demos/microshift/add-microshift-repos.sh
@@ -1,17 +1,9 @@
 #!/bin/bash
 
-echo "Starting..."
-# By default use latest OCP release less 2
-RHOCP_release="4.$(( $(subscription-manager repos | grep rhocp-4 | awk -F - '{print $2}' | awk -F . '{print $2}' | sort -nr | head -n1) - 2 ))"
-
-
 baserelease=$(cat /etc/redhat-release  | awk '{print $6}' | awk -F . '{print $1}')
 basearch=$(arch)
 
 current_dir="$(pwd)"
-
-
-
 
 
 
@@ -70,6 +62,14 @@ sudo composer-cli sources add microshift-local.toml
 
 ###### OPENSHIFT REPO
 
+
+# By default use latest OCP release less 2
+
+subscription-manager repos | grep rhocp-4
+if [ $? = 0 ]
+then
+    RHOCP_release="4.$(( $(subscription-manager repos | grep rhocp-4 | awk -F - '{print $2}' | awk -F . '{print $2}' | sort -nr | head -n1) - 2 ))"
+
     ########
     ######## Download openshift local RPM packages (noarch for python and selinux packages)
     rm -rf openshift-local 2>/dev/null || true
@@ -100,15 +100,68 @@ system = false
 EOF
 
 
-
-
-
 sudo composer-cli sources delete openshift-local 2>/dev/null || true
 sudo composer-cli sources add openshift-local.toml
 
 
 
+else
 
+    if [ ! -d "/etc/osbuild-composer/repositories/" ]
+    then
+        mkdir -p /etc/osbuild-composer/repositories/
+    fi
+
+
+    gpgkey_this=$(cat /usr/share/osbuild-composer/repositories/* | grep gpgkey | head -n 1)
+
+cat <<EOF > additional-repos.json
+{
+    "${basearch}": [
+        {
+            "name": "rhocp",
+            "baseurl": "https://cdn.redhat.com/content/dist/layered/rhel8/${basearch}/rhocp/4.10/os",
+${gpgkey_this}
+            "rhsm": true,
+            "check_gpg": false
+        },
+        {
+            "name": "ansible",
+            "baseurl": "https://cdn.redhat.com/content/dist/layered/rhel8/${basearch}/ansible/2.9/os",
+${gpgkey_this}
+            "rhsm": true,
+            "check_gpg": false
+        }
+    ]
+}
+EOF
+
+
+
+
+ for file in $(ls /usr/share/osbuild-composer/repositories/) 
+ do
+
+ jq -s 'def deepmerge(a;b):
+   reduce b[] as $item (a;
+     reduce ($item | keys_unsorted[]) as $key (.;
+       $item[$key] as $val | ($val | type) as $type | .[$key] = if ($type == "object") then
+         deepmerge({}; [if .[$key] == null then {} else .[$key] end, $val])
+       elif ($type == "array") then
+         (.[$key] + $val | unique)
+       else
+         $val
+       end)
+     );
+   deepmerge({}; .)' /usr/share/osbuild-composer/repositories/$file additional-repos.json > /etc/osbuild-composer/repositories/$file
+
+ done
+
+
+
+
+
+fi
